pyproject.toml
[project]
name = "media-bot-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bencoding>=0.2.6",
    "deluge-client>=1.10.2",
    "imdbpy>=2022.7.9",
    "mysql>=0.0.3",
    "pickledb>=1.3.2",
    "plexapi>=4.17.0",
    "pydantic>=2.11.7",
    "pysocks>=1.7.1",
    "pytelegrambotapi>=4.27.0",
    "python-qbittorrent>=0.4.3",
    "python-telegram-bot[socks]>=22.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "tmdbsimple>=2.9.1",
    "torrent-parser>=0.4.1",
    "transmissionrpc>=0.11",
]
